refactor(rendering): centralize light defaults and improve const-correctness
Extracted hardcoded light constants into light_defaults.h for centralized configuration
- Replaced magic numbers (255, 215, 0.0f) with named constants across brush_manager, ingame_preview_canvas, ingame_preview_renderer, and drawing_options
- Changed TileLocation* to const TileLocation* in tile_renderer and map_layer_drawer APIs for const-correctness
- Changed floor_light_start parameter from size_t to uint32_t in light_buffer and tile_renderer to match actual usage
- Added std::clamp to GUI light setter methods to enforce valid ranges
- Tied draw_floor_shadow to show_shade setting instead of hardcoded true
- Clamped minimum_ambient_light to [0.0f, 1.0f] in drawing_options update
- Moved tile_center_y calculation outside inner loop in light_drawer for performance
- Precomputed light color base values outside tile loop in light_drawer
- Added clarifying comments for OTClient light occlusion and blending behavior
- Removed unnecessary const_cast calls in ingame_preview_renderer and creature_drawer
- Removed preview_local_player light registration path from creature_drawer

+++ source/rendering/core/light_buffer.h
@@ -32,7 +32,7 @@ struct LightBuffer {
 	void AddLight(int pixel_x, int pixel_y, const SpriteLight& light);
 	void AddTileLight(int tile_x, int tile_y, const SpriteLight& light);
 	void AddScreenLight(int screen_x, int screen_y, const RenderView& view, const SpriteLight& light);
-	void SetFieldBrightness(int tile_x, int tile_y, size_t start, uint8_t color = 0);
+	void SetFieldBrightness(int tile_x, int tile_y, uint32_t start, uint8_t color = 0);
 	void Clear();
 
 	[[nodiscard]] bool ContainsTile(int tile_x, int tile_y) const noexcept;
